Extract hotel detail preview from the swipe list item
